# Use a different struct for scrape target serialization