--- a/cmd/otel-allocator/server/bench_test.go
+++ b/cmd/otel-allocator/server/bench_test.go
@@ -198,7 +198,7 @@ func BenchmarkTargetItemsJSONHandler(b *testing.B) {
 		},
 	}
 	for _, tc := range tests {
-		data := makeNTargetItems(*random, tc.numTargets, tc.numLabels)
+		data := makeNTargetJSON(*random, tc.numTargets, tc.numLabels)
 		b.Run(fmt.Sprintf("%d_targets_%d_labels", tc.numTargets, tc.numLabels), func(b *testing.B) {
 			b.ReportAllocs()
 			for i := 0; i < b.N; i++ {
@@ -242,7 +242,7 @@ func makeNCollectorJSON(random rand.Rand, numCollectors, numItems int) map[strin
 	for i := 0; i < numCollectors; i++ {
 		items[randSeq(random, 20)] = collectorJSON{
 			Link: randSeq(random, 120),
-			Jobs: makeNTargetItems(random, numItems, 50),
+			Jobs: makeNTargetJSON(random, numItems, 50),
 		}
 	}
 	return items
@@ -259,6 +259,15 @@ func makeNTargetItems(random rand.Rand, numItems, numLabels int) []*target.Item 
 		))
 	}
 	return items
+}
+
+func makeNTargetJSON(random rand.Rand, numItems, numLabels int) []*targetJSON {
+	items := makeNTargetItems(random, numItems, numLabels)
+	targets := make([]*targetJSON, numItems)
+	for i := 0; i < numItems; i++ {
+		targets[i] = targetJsonFromTargetItem(items[i])
+	}
+	return targets
 }
 
 func makeNNewLabels(random rand.Rand, n int) model.LabelSet {